replica lag banning, multiple bug fixes from HN (#788)
- Fix #787 
- Partial fix for #785 
- Fix #789
- Fix segfault in pg_query `deparse_raw` caused by empty strings
- clippy

+++ pgdog/src/net/messages/data_types/bigint.rs
@@ -2,6 +2,6 @@ use crate::net::DataRow;
 
 impl From<DataRow> for i64 {
     fn from(value: DataRow) -> Self {
-        value.get_int(0, true).unwrap_or(0) as i64
+        value.get_int(0, true).unwrap_or(0)
     }
 }
